[feat] todayTop과 todayBottom 병합
