+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Millie_Android"
         android:usesCleartextTraffic="true">
         <activity
-            android:name=".presentation.todaybottom.TodayActivity"
+            android:name=".presentation.today.TodayActivity"
             android:exported="true"
             android:label="@string/title_activity_today"
             android:theme="@style/Theme.Millie_Android">
@@ -23,9 +23,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".presentation.todayTop.TodayTopActivity"
-            android:exported="true" />
         <activity
             android:name=".presentation.MainActivity"
             android:exported="true"
