+++ app/src/main/java/com/millie/millieshelf/presentation/MainActivity.kt
@@ -3,8 +3,6 @@ package com.millie.millieshelf.presentation
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.millie.millieshelf.R
-import com.millie.millieshelf.data.api.ServicePool
-import retrofit2.await
 
 class MainActivity : AppCompatActivity() {
