+++ app/src/main/java/com/millie/millieshelf/presentation/today/best/TodayBestAdapter.kt
@@ -1,19 +1,18 @@
-package com.millie.millieshelf.presentation.todayTop.best
+package com.millie.millieshelf.presentation.today.best
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.millie.millieshelf.databinding.ItemTodayBestContentBinding
-import com.millie.millieshelf.model.TodayBookmarkModel
 import com.millie.millieshelf.model.response.TodayBest
 import com.millie.millieshelf.util.ItemDiffCallback
 
 class TodayBestAdapter(context: Context) :
     ListAdapter<TodayBest.Data, TodayBestViewHolder>(
         ItemDiffCallback<TodayBest.Data>(
-            onContentsTheSame = {old, new -> old == new},
-            onItemsTheSame = {old, new -> old.id == new.id}
+            onContentsTheSame = { old, new -> old == new },
+            onItemsTheSame = { old, new -> old.id == new.id }
         )
     ) {
     private val inflater by lazy { LayoutInflater.from(context) }
