+++ app/src/main/res/layout/activity_today.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.todaybottom.TodayActivity">
+        tools:context=".presentation.today.TodayActivity">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fcv_today"
