+++ app/src/main/res/values/strings.xml
@@ -12,7 +12,6 @@
     <string name="TEMP">temp</string>
     <string name="BEST_PERCENT_TIME">%1$d%% | %2$d분</string>
 
-    <string name="BINDING_ERROR">BINDING ERROR</string>
     <string name="title_activity_today">TodayActivity</string>
     <!-- Strings used for fragments for navigation -->
     <string name="first_fragment_label">First Fragment</string>
